refactor: move large realistic_repo fixture to separate module
Move 1200+ lines of test fixture file content from conftest.py to a
dedicated tests/fixtures/realistic_repo_data.py module:

- Created tests/fixtures/ package with realistic_repo_data.py
- Extracted REALISTIC_REPO_FILES list with all 10 file contents
- Added create_realistic_repo() helper function
- Simplified realistic_repo fixture to use the helper
- Reduced conftest.py from 1548 to 327 lines (79% reduction)

No behavior changes - fixture creates identical test repository.

Implements #273
