+++ tests/fixtures/__init__.py
@@ -1,0 +1,8 @@
+"""Test fixtures module."""
+
+from tests.fixtures.realistic_repo_data import (
+    REALISTIC_REPO_FILES,
+    create_realistic_repo,
+)
+
+__all__ = ["REALISTIC_REPO_FILES", "create_realistic_repo"]
